pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "efficient-research"
version = "0.1.0"
description = "Multi-agent research + workshop writing system"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["orchestrator*", "agents*", "tools*", "eval*", "schemas*"]
